Move decodePublicKeys into a shared helper
The keymanager loads keys from the flag and the key file too, so decoding and
deduping hex public keys is no longer specific to the URL source. Adds a pubkey
alias for the raw BLS key type the rest of the change is built on.
