+++ validator/keymanager/remote-web3signer/BUILD.bazel
@@ -3,6 +3,7 @@ load("@prysm//tools/go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
+        "helper.go",
         "key_source_file.go",
         "key_source_url.go",
         "keymanager.go",
@@ -43,6 +44,7 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
+        "helper_test.go",
         "key_source_file_test.go",
         "key_source_url_test.go",
         "keymanager_test.go",
